move auth headers to utils
