+++ train.py
@@ -8,7 +8,12 @@ import json
 import uuid
 import boto3
 import os
-from module.utils import create_json_response, require_env, load_sentry
+from module.utils import (
+    create_json_response,
+    require_env,
+    load_sentry,
+    get_auth_headers,
+)
 import datetime
 import base64
 from module.logger import get_logger
@@ -29,21 +34,6 @@ log.info(f"using queue {queue_url}")
 # todo endpoint_url="http://localhost:8000") for localstack
 dynamodb = boto3.resource("dynamodb", region_name=aws_region)
 job_table = dynamodb.Table(JOBS_TABLE_NAME)
-
-
-def get_auth_headers(event) -> Dict[str, str]:
-    authorization = (
-        event["headers"]["Authorization"]
-        if "Authorization" in event["headers"]
-        else None
-    )
-    origin = event["headers"]["Origin"] if "Origin" in event["headers"] else None
-    headers = {}
-    if authorization:
-        headers["Authorization"] = authorization
-    if origin:
-        headers["custom-external-origin"] = origin
-    return headers
 
 
 def handler(event, context):
